actor encode decode params result, fix empty cbor
