+++ core/vm/runtime/runtime.hpp
@@ -14,7 +14,7 @@
 #include "primitives/chain_epoch/chain_epoch.hpp"
 #include "primitives/sector/sector.hpp"
 #include "storage/ipfs/datastore.hpp"
-#include "vm/actor/actor.hpp"
+#include "vm/actor/actor_encoding.hpp"
 #include "vm/exit_code/exit_code.hpp"
 #include "vm/indices/indices.hpp"
 #include "vm/message/message.hpp"
@@ -41,27 +41,6 @@ namespace fc::vm::runtime {
   using primitives::sector::SealVerifyInfo;
   using storage::ipfs::IpfsDatastore;
   using Serialization = Buffer;
-
-  /// Encode actor params, raises appropriate error
-  template <typename T>
-  outcome::result<MethodParams> encodeActorParams(const T &params) {
-    auto maybe_bytes = codec::cbor::encode(params);
-    if (!maybe_bytes) {
-      return VMExitCode::ENCODE_ACTOR_PARAMS_ERROR;
-    }
-    return MethodParams{maybe_bytes.value()};
-  }
-
-  template <typename T>
-  auto decodeActorReturn(const InvocationOutput &result) {
-    return codec::cbor::decode<T>(result.return_value);
-  }
-
-  template <typename T>
-  outcome::result<InvocationOutput> encodeActorReturn(const T &result) {
-    OUTCOME_TRY(encoded, codec::cbor::encode(result));
-    return InvocationOutput{Buffer{encoded}};
-  }
 
   /**
    * @class Runtime is the VM's internal runtime object exposed to actors
@@ -180,9 +159,9 @@ namespace fc::vm::runtime {
     outcome::result<typename M::Result> sendM(const Address &address,
                                               const typename M::Params &params,
                                               TokenAmount value) {
-      OUTCOME_TRY(params2, encodeActorParams(params));
+      OUTCOME_TRY(params2, actor::encodeActorParams(params));
       OUTCOME_TRY(result, send(address, M::Number, params2, value));
-      return decodeActorReturn<typename M::Result>(result);
+      return actor::decodeActorReturn<typename M::Result>(result);
     }
 
     /// Send funds
@@ -206,7 +185,7 @@ namespace fc::vm::runtime {
                               MethodNumber method_number,
                               const P &params,
                               BigInt value) {
-      OUTCOME_TRY(params2, encodeActorParams(params));
+      OUTCOME_TRY(params2, actor::encodeActorParams(params));
       return sendR<R>(to_address, method_number, MethodParams{params2}, value);
     }
 
@@ -216,7 +195,7 @@ namespace fc::vm::runtime {
                                             MethodNumber method_number,
                                             const P &params,
                                             BigInt value) {
-      OUTCOME_TRY(params2, encodeActorParams(params));
+      OUTCOME_TRY(params2, actor::encodeActorParams(params));
       return send(to_address, method_number, MethodParams{params2}, value);
     }
